fix(YALB-261): Remove attempt to disable cache
After digging, there will be quite a few times where even if caching is
disabled, you'll still need to clear drupal so you can get to the
singletons created throughout.  As such, I decided not to touch caching
from a programatic level.

If you wish to do this still, you'll need to modify the settings.php
file to accomplish this.