fix(YALB-261): Add twig configuration and disclaimer
Twig files are tricky to debug through xdebug by default.  I was able to
get this working in my own editor, but not necessarily in vscode due to
not being able to provide breakpoints inside of the twig files
themselves, devel_breakpoint() didn't yield the same results as it did
in neovim, and attempting to debug the generated php files for the twig
output did not trigger.

If anyone that uses vscode more would like to dive into this, that would
be great.