fix(YALB-261): Add VSCode docs to open folder
A very important piece could hinder VSCode users from following the
documentation, and that was that in my testing, I used the terminal to
open vscode from the project root.  Doing this allowed paths to be found
relative to that root, like the .vscode folder that is in the
instructions.

The VSCode instructions have been updated to include this bit of
information so that the relative directories are recognized by the
application.