fix(YALB-1123): Cache invalidated on alert save
When saving an alert, not all pages that had the alert rendered were
becoming invalidated, resulting in pages displaying the previous alert
until expiration.  Now, this will be invalidated each time the settings
are saved.

Thanks to Marc Berger for suggesting to look at a similar fix he made to
another part of the application.

Co-authored-by: Marc Berger <codechefmarc@users.noreply.github.com>