fix(YALB-1195): node variables used
For events, content is not being set, probably due to layout builder.
Node, however, does have the values we need and so this fixes the
variables passed on to the control to use node variables instead.

I believe in the future, this will be changed, but for current
functionality, this is needed.