fix(YALB-1195): all day now detected
One item is missing from the node elements that smart date tracks, and
that is whether the 'All day' option is selected.  As a way around this,
I noticed that when all day is selected for one day, the duration
variable that is present is set to 1439 minutes.  I was able to use this
as a constant of sorts in order to calculate the modulus of the
duration, telling us whether it was a full day.

For 'all day' events spanning multiple days, this broke due to other
full days being 1440 minutes in length.

- 1 day:  1439 minutes
- 2 days: 1439 + 1440 minutes (2879)
- 3 days: 1439 + 1440 + 1440 minutes (4319)

As such, I created a delta that would be the integer portion of the
modulus.  So the delta would be the following:

- 1 day:  1
- 2 days: 2
- 3 days: 3

This allows you to determine still whether it is an all day selection or
not.

Note that this should all be outside of the twig file, but that might be
a different ticket.