fix(YALB-1195): field_event_format can pass data
When testing the output from my previous changes, I noticed that there
was an issue with how the field_event_format was being used.  It was
using the value method, which I understand to take the first element's
value, but in the case of that field, it's possible to have multiple
entries selected (online and in person).  When I attempted to change
this to include the list, the component then refused to work, due to it
checking if there were more than one item, and if there wasn't it did
not reference the 0th element like it does if there were multiple.

The other issue I ran into was that it was passing the backend value to
the component, not the translated label.  I unfortunatelly could not
find the label in the node information to retrieve, so I created my own
list of possibilities for now until we can gain access to this.  You'll
notice that there are two more items included: "In-person" and "Online",
which seem redundant, but when looking at the published
component-library-twig, storybook was not initially displaying the
location due to these not being set.  Once that was there, it properly
displayed on initial load.

This is a please fix me case; a lot of this probably needs to go to the
backend.