feat(YALB-1197): regex constrains soundcloud iframe
This new regex constrains the allowable url from the embed for
SoundCloud. It also targets different matches for re-use in the template
output.

While we are excluding playlists, we need to capture playlists here so
we can later validate it to give a better error message stating we will
only accept tracks.