feat(YALB-971): add git checkout script
This creates a script out of the need to be able to switch branches of
any of the repos at any time.  The new script, local-git-checkout.sh,
allows you to switch any sub-repo you wish to the version you want,
while exiting should there be uncommitted changes so you can take care
of it first and not lose any work.

Note, at this point, I have not tested how well it actually functions on
a drupal site side, only from ensuring it accuately switches branches.