feat(YALB-1197): default build uses embed_wrapper
This changes the base build function to default to the embed_wrapper
theme, making the twig file renderable.  It passes the old array as an
embedSource through as a fallback should the twig file not be able to
handle the data.

The url seemed to be an important part of the current
component-library-twig embed component, so I forced this in the output.
As such, each element has it's own way of constructing a valid URL, so a
base getUrl was needed, which allows subclasses to override.

Co-authored-by: nJim <jamesvomero@gmail.com>