fix(YALB-971): force composer for atomic
When setting up a yalesites project, composer does not use the name
'origin', but instead uses a remote name of 'composer'.  This caused
some issues when attepmting to switch branches on atomic, as there would
be two tracking branches, and it didn't know which to take.  This now
passes the 'composer' remote only for atomic and passes it through.