fix(YALB-1333): source correct location of say.sh
While creating say.sh, I forgot to reference the proper location of it
in the local review scripts, resulting in error messages when attempting
to write messages through the say command.  This now references
correctly, as well as checking for the file so we can know when there is
a problem.