feat(YALB-1233): Update view filters with Chosen
While this initially looks like the right thing to do, the result is
that the two fields do not show at all.  I'm unsure if I'm doing
something incorrectly, or if a specific template needs to be created to
accomodate these?  If so, how is the rest of Drupal able to use Chosen
without these changes?  I'll still be looking for answers, but thought
I'd commit here and possibly squash this once I get through this
blocker.