feat(YALB-1233): ys_views_basic can pull all tags
It's possible that chosen cannot be used in the same way that the drupal
autocomplete can, and so we could need an options array to handle it.
This was made as a way to have this functionality by loading multiple
taxonomy_terms and returning a hash of ids/labels to them.