fix(YALB-1233): Copy allow-plugins to fix CI
CI seems to be erroring, and looking at the script, it changes into the
profile directoy and attempts to use composer.  Because all of our
allowed plugins are defined in the main repo, I could see why this would
error on us.  I'm attmempting to add it in the hopes it'll be happy,
even if it's redundant.