fix(YALB-1233): Minimal revert of previous commit
While the linting and testing were successful by placing the info that
was in the main composer.json into the profile's, the build to pantheon
seems to be failing.  Looking at the error, it's pointing to a file path
that really shouldn't exist.  This made me think I was a bit too heavy
handed on my change, so I'm rolling it back and only keeping the
allow-plugins for composer/installers since that was the original
complaint in the hopes that the build succeeds.