fix(YALB-1394): fix header level for accordion items
The original implementation disregarded the accordion_items in the
component library and rendered them in atomic.  This was causing a
missing h2 level when no accordion title was present.

To fix this, we preprocess the paragraph__accordion_item to inject a new
attribute that will hold the heading_level that should be used,
defaulting to 3.