fix(YALB-1394): pass correct field_heading element
The video element was passing content.field_heading to the
video__heading variable.  This was causing evaluations of whether the
video__heading was empty to result in false due to it being an array I
beleive.  Changing this to always attempt to pass the 0th element now
causes it to function as expected.