feat(YALB-1370): Add tokenizer and HTML filter
Debugging search_api led me to a TextNode object that seems to be
created by Stemmer.  The instructions say that a much better experience
is had if you enable the tokenizer first, which we did not have.  So
stripping the HTML with the tokenizer results in proper search occuring.

Before this change, not only could you search for breadcrumb, but any
menu item, such as home, about, and more.  You'll still be able to
search those pages from their title, but not every page needs to be
returned because they have a top menu item of home.