fix(YALB-1518): proper termId is selected
The current issue is that when a view is accessed that was previously
created using the Drupal autocomplete results in a truthy value (a 1)
being placed for the ID, so that taxononmy term is used for all views
matching this.  This fix properly evaluates whether the term is an
object (the Drupal autocomplete) or an id (the chosen autocomplete) and
has been extracted to its own private method to start to separate
concerns.