feat(YALB-1472): enable area limits to imagemagick
Even with the file limit, it still bases the memory on the area of the
image (width * height * bits needed by imagemagick).  By adding the area
limit, if the width * height gets over 20 megapixels, it utilizes disk
over memory.  This does not mean that it'll load every time, but it now
gives a chance for all of the images to finish.