feat(YALB-1472): memory and map needed
For a successful image generation, memory/map limits are still needed.
While having the area tells it when to use disk space to map, the
mapping must be filling up and erroring still.  While this does not
ensure an image every time, it has usually worked on the 2nd try.