fix(YALB-1472): limit images to 4K resolution
Given what was said by Pantheon, and the calculator provided, I thought
we'd try to see how images would do limiting to the 4k resolution of
3840x2160 while leaving in our imagemagick limits in case somehow an
image gets on the server that would cause it to error.  This will give
larger images a chance to load should they ever get on there while
limiting the amount of memory needed for uploaded images.