feat(YALB-1588): add export creation script
This is a start of a script to create new versions of the content
export.

The idea is that you would clone the yalesites-starterkit repository to
a directory, and then run this script, passing the directory to that
repository.  Provided environment variables were set, it will:
  - Create the export remotely on the yalesites server
  - Retrieve the zip file and delete it from the server
  - Unzip it to the repository location
  - Create a commit with a timestamp
  - Delete the original zip file created

Environment variables needed:
  - PANTHEON_NAME: The terminus site name to connect to
  - SFTP_PORT: the port to use when SFTPing
  - SFTP_URI: the username@location to SFTP to

Usage:
  PANTHEON_NAME=ys-starterkit.dev SFTP_PORT=2222 \
  SFTP_URI="uritothelocation" ./scripts/local/starterkit/create-export \
  /code/yalesites-starterkit

  or preset the environment variables and just call the script with the
  destination repo.