fix(YALB-1593): assign unique dom_id to views
In Drupal, there is a mechanism that attemps to re-use pagers for views,
usually because they are not used multiple times in a page.  For those
times they are, you can use the dom_id as a way to uniquely identify a
view such that the pager will be unique.

In our implementation, the dom_id of our view is never set, causing this
to be skipped, and the pager ID to retain a value of 0, making it
re-used across the page rendered.

This fix assigns a unique dom_id to the view before it is built so
that it will generate a unique pager instance per view.