YALB-1472: Bug: Image generation failures (#423)
* feat(YALB-1472): enable area limits to imagemagick

Even with the file limit, it still bases the memory on the area of the
image (width * height * bits needed by imagemagick).  By adding the area
limit, if the width * height gets over 20 megapixels, it utilizes disk
over memory.  This does not mean that it'll load every time, but it now
gives a chance for all of the images to finish.

* feat(YALB-1472): memory and map needed

For a successful image generation, memory/map limits are still needed.
While having the area tells it when to use disk space to map, the
mapping must be filling up and erroring still.  While this does not
ensure an image every time, it has usually worked on the 2nd try.

* fix(YALB-1472): limit images to 4K resolution

Given what was said by Pantheon, and the calculator provided, I thought
we'd try to see how images would do limiting to the 4k resolution of
3840x2160 while leaving in our imagemagick limits in case somehow an
image gets on the server that would cause it to error.  This will give
larger images a chance to load should they ever get on there while
limiting the amount of memory needed for uploaded images.

* feat(YALB-1472): add image conversion scripts

In order for images to load properly, we made the decision to limit the
width and height to a 4k display.  There are sites out there that
already have these larger images, and will not get resized.  These are
scripts to assist in handling this.

- calculateMB.sh: returns the amount of memory an image will use to
  convert to our currently highest resolution
- get_images.sh: retrieves all files (images) under the /files directory
  of a Pantheon site
- put_images.sh: places all files (images) under a local directory to
  a directory on a Pantheon site
- fix_images.sh: Evaluates all jpgs inside of a directory.  If they are
  above the threshold that we desire of a width/height, it attempts to
  resize it in place.
- doItAll.sh: My attempt to have one script that will retrieve images,
  convert them, and reupload them in one script (still a WIP)