fix(YALB-1588): use rev-parse for git detection
The path given for the starterkit output was the root directory of a git
repository, so we were testing for the `.git` directory to be present.
Now that we are embedding a directory underneath, we no longer have the
.git directory on the given directory.  Instead, we now use `git
rev-parse --show-toplevel` to attempt to get the path to the git root.
If it doesn't exist, it should continue on wihout committing.