fix: add conditional logic to drush deploy quicksilver script
`drush deploy` that runs in CI was getting no output because it was first running on the platform from Quicksilver.
This updated script uses Pantheon's workflow API endpoint to check if the sync_code hook was initiated by our machine user or not.
This commit also renames this file and references in pantheon.yml, since it is now doing more than just a config import.