fix(YALB-1588): remove old files before unzipping
When exporting, there may be a case where we have an asset or node that
no longer should exist.  Do accomodate this, we remove all files in the
directory so we can unzip them fresh.

We add some safety input controls here because we do not want to delete
folders that would harm their system, like if someone accidentally put /
or ~.  So we now have a prompt to ensure they want to delete all files
within the directory specified so that they can choose to cancel at that
time.