feat(YALB-1588): use latest version of starterkit
We can always modify this, but it was recommended to get the latest
version of the starterkit whenever we do the import.  We can always
revert this if we find a case that would make us reconsider using
latest, but I believe it was discussed and agreed upon that latest is
appropriate.