YALB-1574: Branding: Header Variation Backend Front Page Alias Fix (#491)
* fix(YALB-1574): front page alias shows header

This attempts to fix a bug that was found with the header
implementation.  isFrontPage does not seem to take into account aliases.

We get around this by getting the node id of the page, the alias of the
page, and the current front page that is set in site settings.  We then
determine if it is the front page if the location set in the site
settings matches either the node id path or the alias path, or if
isFrontPage thinks it's a front page.

References: https://www.drupal.org/project/drupal/issues/1503146