fix(YALB-1602): paragraph references cloned (#496)
Quick Node Clone does a great job of cloning, however, when it attempts
to clone a node with layout builder, it does not correctly clone the
inline blocks and paragraph references we use.  This is a more global
issue, and a patch has been worked on to fix this.  This patch seems to
fix our issue and sub-items are properly cloned and editable.