fix(YALB-1685): update bundle to always be main
In this bug, we finally found the issue to be that these nodes are in a
different bundle, missing the CTA field all together.  If they were
main, it would display as it should as they are defined.  Since we do
not have control over which field definitions get loaded, and it seems
to use the bundle to load them, this updates any that have been
mis-labelled to main so that nodes could be edited again.

@Co-authored-by Berger, Marc <@codechefmarc>