revert(YALB-1685): extra menu items on node sidebar
This reverts the extra menu items on the node sidebar, which is causing
the error due to the bundle being one such that the field definitions do
not contain the CTA mega menu top level title field.  It also reverts
the possible fix to update the bundle to main on cache refresh.

Now, the only way to add this field data is to do so via the Manage Main
Menu view, which works as intended.