fix(YALB-1663): deleting default section does not error
There was a problem where a user deleted the default content section.
Upon saving this, they were unable to go back in to modify the page in
layout builder.  This was due to an access check issue in
section_library.  This patch fixes this so that the user could edit the
page in layout builder again.