YALB-1663: User getting error page when clicking "edit layout and content" in top menu (#530)
* fix(YALB-1663): deleting default section does not error

There was a problem where a user deleted the default content section.
Upon saving this, they were unable to go back in to modify the page in
layout builder.  This was due to an access check issue in
section_library.  This patch fixes this so that the user could edit the
page in layout builder again.

* feat(YALB-1663): allow "Add Section" after Title and Metadata

layout_builder_lock was disallowing sections to be added after the Title
and Metadata section of a Page, or the bottom content section of a Profile.
Since we are now allowing them to remove the default content section, they
will need a way to add a new section, which this provides.  "Add Section" is
now displayed below in these cases.