feat(YALB-1687): create very crude form selection
This is a first step in order to grasp a handle on how controllers
interact with forms in order to display the selections needed to create
a template.

Right now this just goes to the content type addition, but later it'll
use the single content sync api to generate a new node via an import.