feat(YALB-1687): use fallback image if needed
If the image used in the template does not exist, we currently have no
way to bundle this with the zip without doing some unzipping/rezipping,
so this is an easier way to ensure all images are not broken.  Instead,
if the image does not exist, we use a fallback image to represent it.