fix(YALB-1685): Error when going to "Manage Settings" on certain pages (#525) (#542)
* fix(YALB-1685): update bundle to always be main

In this bug, we finally found the issue to be that these nodes are in a
different bundle, missing the CTA field all together.  If they were
main, it would display as it should as they are defined.  Since we do
not have control over which field definitions get loaded, and it seems
to use the bundle to load them, this updates any that have been
mis-labelled to main so that nodes could be edited again.

@Co-authored-by Berger, Marc <@codechefmarc>

* fix(YALB-1685): do not directly access the database

While the previous implementation worked, it felt strange directly
accessing the database.  This goes through Drupal objects to accomplish
the same result, which feels a bit better in case there is any hooks or
background processing that Drupal does in these cases.

* revert(YALB-1685): extra menu items on node sidebar

This reverts the extra menu items on the node sidebar, which is causing
the error due to the bundle being one such that the field definitions do
not contain the CTA mega menu top level title field.  It also reverts
the possible fix to update the bundle to main on cache refresh.

Now, the only way to add this field data is to do so via the Manage Main
Menu view, which works as intended.

* fix(YALB-1667): find parent page title for basic menu

This sets up component library twig to be able to retrieve the node
title later if criteria is met that makes it use node_title.

This attempts to not touch the mega menu logic, hence the new variable.

See component-library-twig/components/02-molecules/menu/_yds-menu-item.twig