fix(YALB-1696): guard against empty block_form
Code was added in YALB-959 that attached javascript to specific block
types:

* tabs
* quick_links
* media_grids
* gallery

It referenced a block_form array that tells us which block type we are
currently attempting to use.  In some cases, though, like the heading
block, this block_form array does not exist, resulting in a background
error.  This makes it look as if the user cannot modify the block.

We fix this by creating a guard clause; when the block_form does not
exist, we bypass the attachment of this javascript.