fix(YALB-1696): get block type from buildInfo
While block_form doesn't always exist, it seems that getting the
blockInfo and looking at the args can tell you which block type we are
dealing with.  I'm hoping there's a more elegant way of doing this, but
it does seem to be working for existing and new block elements.