fix(YALB-1696): fallback js error to submit button
In fixing the block_form issue, I noticed that when testing the
functionality of YALB-959, there was an issue where paragraph items that
weren't expanded couldn't display the error message to the user, and
would error.  It would not hinder the submissions of the form, so it
would result in users being able to submit with one tab, for instance.

This fix checks if the input that was found is visible, and if not, it
appends the message to the submit button that we know is there, thus
letting the form continue to submit or stop.