fix: REL-100 Hotfix: add guard clauses to bypass <nolink> (#543)
* fix(rel-100-submenu-display-null-node-fix): guard against nolink

In a few sites, they used top level links pointing to <nolink>, which
seems to respond that it is routable, but has no node associated.  We
now guard each variable to check for this so that it can gracefully
fallback in these cases and possibly others we haven't covered yet.