feat(YALB-1687): support zip files
Spagetti code to the max, but this is my first attempt to make content
import work.

This attempts to modify the zip files so that a new node can be created
each time.  It will then wrap the batch creation that Single Content
Sync makes so that we can redirect to the edit page afterward instead of
back to the importer.