YALB-520: Links: Dynamically apply link treatments (#467)
* fix(YALB-520): fix embed source manager array

EmbedSourceManager was not using the correct array to track the sources
already used.  This now references the instances array.

* feat(YALB-520): add twig Url detector

In order for dynamic link treatments to properly function, we need a 
Twig-based Url detector that can be utilized.  This adds that function.

This will take a Url and determine what type of link it is so that it
can be properly treated.