YSP-324: Deprecation warning in smart_date causes axe to find issue (#568)
While fixed, smart_date 4.0.3 is not released yet, and we don't want to
point to dev.  This patch adds a dateFormatter to the class using it
instead of relying on the parent class to provide this, causing the
deprecation.

Once 4.0.3 is released or we decide to bump to 4.1.x release, we can
remove this.