fix(YSP-355): External view results don't have external icon (#581)
* fix(YSP-355): add field_external_source to config

While field_external_source was defined, it was not included in the
content portion of the config, making it not available to the profile
directory.  With this change, it now passes this data on so that link
treatments can properly be made.

* fix(YSP-355): enable external link passing for search results

While event seemed to have field_external_source passing through, the
others did not.  This enables it so that link treatments can be applied.

* feat(YSP-355): Add search API rendered content default view mode for profiles

* fix(YSP-355): Fix errors in log about profile fields

---------

Co-authored-by: Marc Berger <marc@fourkitchens.com>