YSP-355: External view results don't have external icon (#206)
* fix(YSP-355): views now pass external URL for url

It was found that in views, a content type that was really an external
link was not being designated in the views as such.  This was due to not
passing the external link through to the component to render.  This now
renders the external URL over the internal one if it exists.