YALB-1691: Bug: Min/max referenced items in reusable block (#560)
* fix(YALB-1691): capture global block type

By looking in a different place for the block type on a global block, we
can identify the type that it is so we can know whether to append the
block_form.js file.

* feat(YALB-1691): get block type from global add

When adding a global block, the inline arguments are the UUID of the
block, not the block type.  We add checks for this such that when that
happens we go look up the block and get the type it should be.

* fix(YALB-1691): handle multiple inputSelectors

* Handles multiple inputSelectors so that it can capture all of the
  different kinds that may be on the page, and does this as the user
  messes with the form layout (i.e. adding a new item might make the
  original input no longer show).  First one found wins.
* The script was being loaded multiple times, causing the submit button
  to have multiple click events attached, making the messages flicker in
  and out.  Now it attaches once to the buttons.

* fix(YALB-1691): remove gin_lb patch for successful build

I'm unsure what this patch was addressing looking at the commit that
made it.  In any event, the latest dev version of the gin_lb module
totally changes how they address this array, making the patch unusable.

If we find we need it, I can look into an updated patch, but I'm unsure
what "multivalued" items we have, unless it's speaking about the
paragraph items (tab items, etc.), which seem to do just fine without
the patch.