fix(YSP-374): guard against deleted focus header image
If you set your site to use a focus header, and then you go into media
library and delete the image you specified as the focus header image, it
results in the whole site being unreachable.  This guards against an
image that has been deleted in that it won't render the focus image.