fix(YSP-374): protect against header image issues
To mirror that of the footer block plugin, we do similar things to the
focus header image in that we handle the case when we have an array of
ids over one id for the media image, which could cause it to show no
image.  We also guard against the accidental deletion of the set focus
image header from the media library, and let the user know that this has
happened.