YSP-415: Add media text spinning (#213)
* fix(YSP-415): no more spinning text

It turns out that GinLB does this animation on purpose it seems.  So we
target the throbber (spinner) and disable the animation.  We also set
the display to inline so that the text displays in a more desirable
location; otherwise the text can be clipped by the other div elements.

* fix(YSP-415): messages are clearly visible

While removing the spinning text, it causes the text that did appear for
the user to be obfuscated by the delete and edit links that usually
appear above the image.  This ensures that the throbber appears in front
of those, hiding the buttons and allowing the text to be fully shown.