feat(YSP-419): move update to embed module
Since this is concerning the embed module, let's place the update logic
there.

Also, more needed to be done, like changing the file properties to point
to the new image, copying that image into the right spot, and updating
the footer to now use the new identifier.

It's possible I will be moving the footer settings back to ys_core since
it then couples the footer to the embed.