fix(YSP-399): fix localist links
The shadow set for the button was set to the text color instead of what
the current background color was, resulting in a hard to read button
text.  Due to localist defining the color of the button, we copied this
value from them in order to make it look correct.

The other links have our animation effect, however localist was defining
a hover line to appear, which made this look awful.  This forces the
links to not underline and to instead use our underline effect.