fix(YSP-419): Rebrand Twitter to X (#627)
* feat(YSP-419): update embed

* feat(YSP-419): update social links manager

* fix(YSP-419): update twitter data to x-twitter

Update the configuration currently in the twitter config item to
x-twitter config item so that it can still be used going forward.

* feat(YSP-419): move update to embed module

Since this is concerning the embed module, let's place the update logic
there.

Also, more needed to be done, like changing the file properties to point
to the new image, copying that image into the right spot, and updating
the footer to now use the new identifier.

It's possible I will be moving the footer settings back to ys_core since
it then couples the footer to the embed.

* docs(YSP-419): Update twitter reference to X

* fix(YSP-419): remove twitter config reference

Since it is only used on initial configuration import, this is not
needed.

* feat(YSP-419): use new wording for X post

* fix(YSP-419): rename twitter again

* feat(YSP-419): darken embed icons to match X

* feat(YSP-419): replace all embed images

The embeds do not use the location that we made the darken edits, but
instead uses the site's files directory to store these.  This ensures
that these new files are copied (overwriting) to that directory so that
the icon can be referenced correctly in the platform.