fix(YSP-528): add tags to table media view
While users can add tags, there wasn't a usage for them yet in the
system.  This was a way to still convey that entering this can be useful
by giving them a chance to filter via the media table by tags and see
what tags are assigned to a media.

This adds both a tags filter using Drupal's autocomplete, and a tags
column to display tags associated with media.