fix(YSP-525): remove bulk publish/unpublish (#647)
Since we are using content moderation, using the existing bulk
operations of publish and unpublish result in a warning notifying the
user of the use of content moderation, followed by error messages
reinforcing the same thing.  Since this is now not able to be done, we
remove these options from being selected in the bulk operations drop
down.