fix(YSP-538): make pathauto happen later
While the previous commit fixed the error, the URL alias was not updated
when it came time for pathauto to evaluate the parents, resulting in
having to save twice to get the real URL alias.  This bumps it in the
weight so that it's later in the list, allowing the URL alias to update
on the first save.