fix(YSP-538): fix a menu link error on publish (#654)
* fix(YSP-538): make workflow_buttons occur before tokens

The issue here was that the workflow_button actions were not completing
before tokens was attempting to use the menu_tree.  Now that it is
weighted to happen first, the menu item appears in the menu_tree and all
is well again.

* fix(YSP-538): make pathauto happen later

While the previous commit fixed the error, the URL alias was not updated
when it came time for pathauto to evaluate the parents, resulting in
having to save twice to get the real URL alias.  This bumps it in the
weight so that it's later in the list, allowing the URL alias to update
on the first save.