fix(YSP-570): change limit behavior on display change
The original functionality that was not working was that when display
changes, the following should happen:

1. Limit input should not show when 'all' display selected
2. Limit input label should be 'Items' when 'limit' display selected
3. Limit input label should be 'Items Per Page' when 'pager' display
   selected

Try as I might, there seems to be issues when you have two elements, one
with a #states set for visibility, and another set to a callback that
changes the other element with the #states array.  As such, I decided to
move this logic to javascript since we already had a JS file associated
and it now takes care of this very quickly.