fix(YSP-535): let cta-banner overlap empty div
In this specific case, there exists an empty div where the content
moderation block renders.  This empty div does not allow the action
banner to appear flush with the header.  This specific fix attempts to
make the specificity of this win such that any layout--banner's
layout--region who first child is a div that contains no class, id,
name, or style (essentially empty)'s sibling that is a .cta-banner have
no top margin.