fix(YSP-576): add permissions for new blocks
While looking at another issue, we noticed that the permissions for the
new blocks were not set.  While this doesn't hinder users from using
them due to relying on Layout Builder Restrictions, it is our convention
to still enable them in case another module relies on these.  This
enables them to mimic other blocks in the past.