feat(YSP-580): add drush command to update profile defaults
This adds a drush command that we can choose to run to reset profiles
with missing dial data to the original look and feel from before the
v1.4.0 release date.

It pulls all profile nodes that have been last updated before the
v1.4.0 date and if there exists no values for image_orientation,
image_style, or image_alignment, it will set it to values that would
exhibit the old look of the profiles.