fix(YSP-551): fix reusable block errors
It turns out cloned blocks were actually still the same block, but with
a different revision.  Because of this, when you set it up to be
reusable, it would actually fail the save, making the block in an
inconsistent state as we then try to create a new block to replace the
now reusable (we thought) block.  This attempts to save in the regular
way, but in a failure, it assumes it's in this state and breaks off to
duplicate itself as a new block so that it can become reusable.