fix(YSP-623): prepend missing protocols on _url fields
If a field ends in _url, attempt to see if the value is a valid URL.  If
it is not and it happens to be missing the protocol, prepend https://
such that it has a better chance of being a valid URL.

This is due to localist not checking for http/https protocols in the URL
fields.  We do this as a way to help it become compatible with Drupal
URI object parsing.