fix(YSP-623): allow event_cta to be overwritable
When syncing data, the event_cta was not one of the allowed items to be
overwritten on an update to the data.  Because of this, it was suspected
that errors were occuring due to this bypassing some validation,
creating URLs that don't actually parse well to a URL object in Drupal.
We now allow this to be overwritten on a sync should it change.