feat(YSP-593): set sync source so we recreate layout builder
A byproduct of setting syncSource is that the layout builder items need
to be recreated.  This is a benefit to us I believe in that we don't
need to figure out if it needs updating, we just add a new block with
the new content.

My only worry is that database data is not being deleted when this
happens; I need to look more into this.