fix(remove-exposed-filter-patch): Remove filter patch for now
Jakala relies on a patch that allows multiple views exposed filters to
apply solely on the view that initiated it.  Unfortunately, this
resulted in issues with the media filter not functioning properly and
providing an access denied on filter.  For now, we will remove this for
release while Jakala determines a new patch that could accomodate both
situations.