fix(thumbnail-view-defaults): default selected on no config
When an existing view is modified, the default was that show thumbnail
would be false, when in fact, it should default to true.  When creating
new views, it does this, but for those that had no value at all, it
defaulted to unchecked.

This checks if fieldOptionValue is empty and if so, it uses the default
value of show_thumbnail enabled.