fix(YSP-680): fix deprecation warning on first load
It looks as if there is a deprecation warning on first load regarding
the placement of $cachable and $entity on the
ys_themes_allowed_values_function in ys_themes.  I am attempting to
switch the order of the variables to squelch this deprecation warning on
first load for sites.

The deprecation warning:
Deprecated: Optional parameter $entity declared before required
parameter $cacheable is implicitly treated as a required parameter in
/code/web/profiles/custom/yalesites_profile/modules/custom/ys_themes/ys_themes.module
on line 28