fix(YSP-593): do not overwrite old data each time
Marc has written about this, and it looks as if when an item is no
longer present in the payload that it gets wiped, which we do not want.

So we removed this; more testing needs to be done on the effects of
changing the data and how it updates.