docs(git): speak about making atomic a safe directory
In some cases, vendors were receiving strange errors related to the
atomic directory already existing.  It turned out the real issue was
that git detected strange ownership of it, and would disallow it from
being brought down.

While the first fix is to ensure that yalesites-org/* was marked in
composer as prefer-dist in the config.json, the next is to add a safe
directory to allow this to work again.